# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soltex Global Admin API",
    "version": "0.3.0",
    "description": "Phase C: authentication, second factors, users and audit. Errors are RFC 9457 Problem Details (application/problem+json) with `request_id`; every response carries `X-Request-ID`. State-changing requests need the header `x-requested-with: soltex-admin`."
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-soltex_session",
        "description": "Opaque session token (HttpOnly cookie set by /auth/login)."
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-requested-with",
        "description": "Must be \"soltex-admin\" on POST/PUT/PATCH/DELETE."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ],
              "additionalProperties": false
            }
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "request_id"
        ],
        "additionalProperties": false,
        "description": "RFC 9457 Problem Details"
      },
...
    }
  },
  "paths": {
...
    "/api/v1/auth/passkeys/registration/options": {
      "post": {
        "summary": "Start registering a passkey",
        "tags": [
          "passkeys"
        ],
        "description": "Returns PublicKeyCredentialCreationOptionsJSON. The challenge is bound to this session, valid 5 minutes, single use.",
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "enrollment",
          "stepUpIfEnrolled": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/passkeys/registration/verify": {
      "post": {
        "summary": "Finish registering a passkey",
        "tags": [
          "passkeys"
        ],
        "description": "If this is the first second factor, the session is raised to assurance 2 and recovery codes are returned (once).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60
                  },
                  "response": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "maxLength": 1024
                      },
                      "rawId": {
                        "type": "string",
                        "maxLength": 1024
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "public-key"
                        ]
                      },
                      "response": {
                        "type": "object",
                        "properties": {
                          "clientDataJSON": {
                            "type": "string",
                            "maxLength": 8192
                          },
                          "attestationObject": {
                            "type": "string",
                            "maxLength": 65536
                          }
                        },
                        "required": [
                          "clientDataJSON",
                          "attestationObject"
                        ],
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "id",
                      "rawId",
                      "type",
                      "response"
                    ],
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "name",
                  "response"
                ]
              }
            }
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "enrollment",
          "stepUpIfEnrolled": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "next": {
                      "type": "string",
                      "enum": [
                        "authenticated",
                        "mfa_required",
                        "mfa_enrollment_required"
                      ]
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "owner",
                            "admin",
                            "content_manager",
                            "editor",
                            "seo_specialist",
                            "marketer"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "name",
                        "role"
                      ],
                      "additionalProperties": false
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "content.read",
                          "content.edit",
                          "content.publish",
                          "seo.read",
                          "seo.edit",
                          "seo.publish",
                          "redirects.read",
                          "redirects.manage",
                          "media.read",
                          "media.upload",
                          "media.replace",
                          "media.delete",
                          "leads.read",
                          "leads.manage",
                          "leads.export",
                          "users.manage",
                          "settings.manage",
                          "system.read",
                          "system.deploy",
                          "system.rollback",
                          "audit.read"
                        ]
                      }
                    },
                    "session": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "aal": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "expiresAt": {
                          "type": "string"
                        },
                        "stepUpValidUntil": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "aal",
                        "expiresAt",
                        "stepUpValidUntil"
                      ],
                      "additionalProperties": false
                    },
                    "mfa": {
                      "type": "object",
                      "properties": {
                        "requirement": {
                          "type": "string",
                          "enum": [
                            "passkey",
                            "any",
                            "none"
                          ]
                        },
                        "requirementMet": {
                          "type": "boolean"
                        },
                        "enrolled": {
                          "type": "boolean"
                        },
                        "passkeys": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "totp": {
                          "type": "boolean"
                        },
                        "recoveryCodesLeft": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "requirement",
                        "requirementMet",
                        "enrolled",
                        "passkeys",
                        "totp",
                        "recoveryCodesLeft"
                      ],
                      "additionalProperties": false
                    },
                    "recoveryCodes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "next",
                    "user",
                    "permissions",
                    "session",
                    "mfa"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/passkeys/authentication/options": {
      "post": {
        "summary": "Start a passkey check (sign-in second factor or step-up)",
        "tags": [
          "passkeys"
        ],
        "description": "Returns PublicKeyCredentialRequestOptionsJSON for the passkeys of the signed-in user.",
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "pending"
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/passkeys/authentication/verify": {
      "post": {
        "summary": "Finish a passkey check",
        "tags": [
          "passkeys"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "response": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "maxLength": 1024
                      },
                      "rawId": {
                        "type": "string",
                        "maxLength": 1024
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "public-key"
                        ]
                      },
                      "response": {
                        "type": "object",
                        "properties": {
                          "clientDataJSON": {
                            "type": "string",
                            "maxLength": 8192
                          },
                          "authenticatorData": {
                            "type": "string",
                            "maxLength": 8192
                          },
                          "signature": {
                            "type": "string",
                            "maxLength": 2048
                          }
                        },
                        "required": [
                          "clientDataJSON",
                          "authenticatorData",
                          "signature"
                        ],
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "id",
                      "rawId",
                      "type",
                      "response"
                    ],
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "response"
                ]
              }
            }
          }
        },
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "pending"
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/passkeys": {
      "get": {
        "summary": "My passkeys",
        "tags": [
          "passkeys"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "x-soltex-access": {
          "level": "enrollment"
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "name": {
                            "type": "string"
                          },
                          "deviceType": {
                            "type": "string"
                          },
                          "backedUp": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "lastUsedAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "deviceType",
                          "backedUp",
                          "createdAt",
                          "lastUsedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/passkeys/{id}": {
      "delete": {
        "summary": "Remove one of my passkeys (step-up required)",
        "tags": [
          "passkeys"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "x-soltex-access": {
          "level": "full",
          "stepUp": true
        },
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "security": [
    {
      "session": [],
      "csrf": []
    }
  ],
...
}